+++ src/components/Buttons/styles.module.css
@@ -1,0 +1,3 @@
+.autho-buttons {
+  background-color: rgba(46, 46, 46, 0.8);
+}
